feat(adk): improve summarization code
